Add files via upload
This is an updated version of the KDS that is closer to the vision we imagined. This version has the features needed and connects to the Flask backend. You'll notice that there are two files. The index.html file is for new or pending orders. History.html shows completed orders.  

Compatibility note: The KDS expects variables like customer_name, items, status, etc... There's a good chance your Flask API uses different names or routes, so update names accordingly before running.